# docs/api/openapi_v001.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Managed Identity Wallets API",
    "description": "Managed Identity Wallets API",
    "termsOfService": "https://www.eclipse.org/legal/termsofuse.php",
    "contact": {
      "name": "Eclipse Tractus-X",
      "url": "https://projects.eclipse.org/projects/automotive.tractusx",
      "email": "tractusx-dev@eclipse.org"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://github.com/eclipse-tractusx/managed-identity-wallets/blob/develop/LICENSE"
    },
    "version": "0.0.1"
  },
  "security": [
    {"Authenticate using access_token": []},
    {"sts_token": []}
  ],
  "servers": [],
  "paths": {
...
    "/api/presentations": {
      "post": {
        "tags": [
          "Verifiable Presentations - Generation"
        ],
        "summary": "Create Verifiable Presentation",
        "description": "Permission: **update_wallets** OR **update_wallet** (The BPN of the issuer of the Verifiable Presentation must equal to BPN of caller) \n\n Create a verifiable presentation from a list of verifiable credentials, signed by the holder",
        "operationId": "createPresentation",
        "parameters": [
          {
            "name": "audience",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "asJwt",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "object"
                }
              },
              "example": {
                "verifiableCredentials": [
                  {
                    "id": "did:web:localhost:BPNL000000000000#f73e3631-ba87-4a03-bea3-b28700056879",
                    "@context": [
                      "https://www.w3.org/2018/credentials/v1",
                      "https://catenax-ng.github.io/product-core-schemas/businessPartnerData.json",
                      "https://w3id.org/security/suites/jws-2020/v1"
                    ],
                    "type": [
                      "VerifiableCredential",
                      "BpnCredential"
                    ],
                    "issuer": "did:web:localhost:BPNL000000000000",
                    "expirationDate": "2024-12-31T18:30:00Z",
                    "issuanceDate": "2023-07-19T09:11:34Z",
                    "credentialSubject": [
                      {
                        "bpn": "BPNL000000000000",
                        "id": "did:web:localhost:BPNL000000000000",
                        "type": "BpnCredential"
                      }
                    ],
                    "proof": {
                      "created": "2023-07-19T09:11:39Z",
                      "jws": "eyJhbGciOiJFZERTQSJ9..fdn2qU85auOltdHDLdHI7sJVV1ZPdftpiXd_ndXN0dFgSDWiIrScdD03wtvKLq_H-shQWfh2RYeMmrlEzAhfDw",
                      "proofPurpose": "proofPurpose",
                      "type": "JsonWebSignature2020",
                      "verificationMethod": "did:web:localhost:BPNL000000000000#"
                    }
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Verifiable Presentation",
            "content": {
              "application/json": {
                "examples": {
                  "VP as Json-LD": {
                    "description": "VP as Json-LD",
                    "value": {
                      "vp": {
                        "@context": [
                          "https://www.w3.org/2018/credentials/v1"
                        ],
                        "id": "did:web:localhost:BPNL000000000000#b2e69e47-95f3-48ff-af30-eaaab36431d5",
                        "type": [
                          "VerifiablePresentation"
                        ],
                        "verifiableCredential": [
                          {
                            "id": "did:web:localhost:BPNL000000000000#f73e3631-ba87-4a03-bea3-b28700056879",
                            "@context": [
                              "https://www.w3.org/2018/credentials/v1",
                              "https://catenax-ng.github.io/product-core-schemas/businessPartnerData.json",
                              "https://w3id.org/security/suites/jws-2020/v1"
                            ],
                            "type": [
                              "VerifiableCredential",
                              "BpnCredential"
                            ],
                            "issuer": "did:web:localhost:BPNL000000000000",
                            "expirationDate": "2024-12-31T18:30:00Z",
                            "issuanceDate": "2023-07-19T09:11:34Z",
                            "credentialSubject": [
                              {
                                "bpn": "BPNL000000000000",
                                "id": "did:web:localhost:BPNL000000000000",
                                "type": "BpnCredential"
                              }
                            ],
                            "proof": {
                              "created": "2023-07-19T09:11:39Z",
                              "jws": "eyJhbGciOiJFZERTQSJ9..fdn2qU85auOltdHDLdHI7sJVV1ZPdftpiXd_ndXN0dFgSDWiIrScdD03wtvKLq_H-shQWfh2RYeMmrlEzAhfDw",
                              "proofPurpose": "proofPurpose",
                              "type": "JsonWebSignature2020",
                              "verificationMethod": "did:web:localhost:BPNL000000000000#"
                            }
                          }
                        ]
                      }
                    }
                  },
                  "VP as JWT": {
                    "description": "VP as JWT",
                    "value": {
                      "vp": "eyJraWQiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwIiwidHlwIjoiSldUIiwiYWxnIjoiRWREU0EifQ.eyJzdWIiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwIiwiYXVkIjoic21hcnQiLCJpc3MiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwIiwidnAiOnsiaWQiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwIzM4ZTU2ZTg1LTNkODQtNGEyNS1iZjg1LWFiMjRlYzY4MmMwOSIsInR5cGUiOlsiVmVyaWZpYWJsZVByZXNlbnRhdGlvbiJdLCJAY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvMjAxOC9jcmVkZW50aWFscy92MSJdLCJ2ZXJpZmlhYmxlQ3JlZGVudGlhbCI6eyJAY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvMjAxOC9jcmVkZW50aWFscy92MSIsImh0dHBzOi8vY2F0ZW5heC1uZy5naXRodWIuaW8vcHJvZHVjdC1jb3JlLXNjaGVtYXMvYnVzaW5lc3NQYXJ0bmVyRGF0YS5qc29uIiwiaHR0cHM6Ly93M2lkLm9yZy9zZWN1cml0eS9zdWl0ZXMvandzLTIwMjAvdjEiXSwidHlwZSI6WyJWZXJpZmlhYmxlQ3JlZGVudGlhbCIsIkJwbkNyZWRlbnRpYWwiXSwiaWQiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwI2Y3M2UzNjMxLWJhODctNGEwMy1iZWEzLWIyODcwMDA1Njg3OSIsImlzc3VlciI6ImRpZDp3ZWI6bG9jYWxob3N0OkJQTkwwMDAwMDAwMDAwMDAiLCJpc3N1YW5jZURhdGUiOiIyMDIzLTA3LTE5VDA5OjExOjM0WiIsImV4cGlyYXRpb25EYXRlIjoiMjAyNC0xMi0zMVQxODozMDowMFoiLCJjcmVkZW50aWFsU3ViamVjdCI6eyJpZCI6ImRpZDp3ZWI6bG9jYWxob3N0OkJQTkwwMDAwMDAwMDAwMDAiLCJicG4iOiJCUE5MMDAwMDAwMDAwMDAwIiwidHlwZSI6IkJwbkNyZWRlbnRpYWwifSwicHJvb2YiOnsicHJvb2ZQdXJwb3NlIjoicHJvb2ZQdXJwb3NlIiwidHlwZSI6Ikpzb25XZWJTaWduYXR1cmUyMDIwIiwidmVyaWZpY2F0aW9uTWV0aG9kIjoiZGlkOndlYjpsb2NhbGhvc3Q6QlBOTDAwMDAwMDAwMDAwMCMiLCJjcmVhdGVkIjoiMjAyMy0wNy0xOVQwOToxMTozOVoiLCJqd3MiOiJleUpoYkdjaU9pSkZaRVJUUVNKOS4uZmRuMnFVODVhdU9sdGRIRExkSEk3c0pWVjFaUGRmdHBpWGRfbmRYTjBkRmdTRFdpSXJTY2REMDN3dHZLTHFfSC1zaFFXZmgyUlllTW1ybEV6QWhmRHcifX19LCJleHAiOjE2ODk4MzQ4MDUsImp0aSI6ImIwODYzOWZiLWQ5MWEtNGUwZS1iNmY4LTYzYjdhMzQ1ZTRhZiJ9.80x0AB-OauefdeZfx1cwhitdVKRvCRFeFzYwU73DL7y4w34vu6BdfHWLBGjkwELxkQEoFfiTPOqtuyqhtsyDBg"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "The request could not be completed due to a failed authorization."
          },
          "403": {
            "description": "The request could not be completed due to a forbidden access"
          },
          "404": {
            "description": "Wallet not found with provided identifier",
            "content": {
              "application/json": {
                "examples": {
                  "Wallet not found with provided identifier": {
                    "description": "Wallet not found with provided identifier",
                    "value": {
                      "type": "about:blank",
                      "title": "Error Title",
                      "status": 404,
                      "detail": "Error Details",
                      "instance": "API endpoint",
                      "properties": {
                        "timestamp": 1689762476720
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Any other internal server error",
            "content": {
              "application/json": {
                "examples": {
                  "Internal server error": {
                    "description": "Internal server error",
                    "value": {
                      "type": "about:blank",
                      "title": "Error Title",
                      "status": 500,
                      "detail": "Error Details",
                      "instance": "API endpoint",
                      "properties": {
                        "timestamp": 1689762476720
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authenticate using access_token": []
          }
        ]
      }
    },
    "/api/presentations/validation": {
      "post": {
        "tags": [
          "Verifiable Presentations - Validation"
        ],
        "summary": "Validate Verifiable Presentation",
        "description": "Permission: **view_wallets** OR **view_wallet**  \n\n Validate Verifiable Presentation with all included credentials",
        "operationId": "validatePresentation",
        "parameters": [
          {
            "name": "audience",
            "in": "query",
            "description": "Audience to validate in VP (Only supported in case of JWT formatted VP)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "asJwt",
            "in": "query",
            "description": "Pass true in case of VP is in JWT format",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "withCredentialExpiryDate",
            "in": "query",
            "description": "Check expiry of VC(Only supported in case of JWT formatted VP)",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "object"
                }
              },
              "examples": {
                "VP as JWT": {
                  "description": "VP as JWT",
                  "value": {
                    "vp": "eyJraWQiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwIiwidHlwIjoiSldUIiwiYWxnIjoiRWREU0EifQ.eyJzdWIiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwIiwiYXVkIjoic21hcnQiLCJpc3MiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwIiwidnAiOnsiaWQiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwIzZmZWRmYWRmLWYxNTItNGNjZS05MWQ1LWNjMjhhNzhlMTExMyIsInR5cGUiOlsiVmVyaWZpYWJsZVByZXNlbnRhdGlvbiJdLCJAY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvMjAxOC9jcmVkZW50aWFscy92MSJdLCJ2ZXJpZmlhYmxlQ3JlZGVudGlhbCI6eyJAY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvMjAxOC9jcmVkZW50aWFscy92MSIsImh0dHBzOi8vY2F0ZW5heC1uZy5naXRodWIuaW8vcHJvZHVjdC1jb3JlLXNjaGVtYXMvYnVzaW5lc3NQYXJ0bmVyRGF0YS5qc29uIiwiaHR0cHM6Ly93M2lkLm9yZy9zZWN1cml0eS9zdWl0ZXMvandzLTIwMjAvdjEiXSwidHlwZSI6WyJWZXJpZmlhYmxlQ3JlZGVudGlhbCIsIkJwbkNyZWRlbnRpYWwiXSwiaWQiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwI2Y3M2UzNjMxLWJhODctNGEwMy1iZWEzLWIyODcwMDA1Njg3OSIsImlzc3VlciI6ImRpZDp3ZWI6bG9jYWxob3N0OkJQTkwwMDAwMDAwMDAwMDAiLCJpc3N1YW5jZURhdGUiOiIyMDIzLTA3LTE5VDA5OjExOjM0WiIsImV4cGlyYXRpb25EYXRlIjoiMjAyNC0xMi0zMVQxODozMDowMFoiLCJjcmVkZW50aWFsU3ViamVjdCI6eyJpZCI6ImRpZDp3ZWI6bG9jYWxob3N0OkJQTkwwMDAwMDAwMDAwMDAiLCJicG4iOiJCUE5MMDAwMDAwMDAwMDAwIiwidHlwZSI6IkJwbkNyZWRlbnRpYWwifSwicHJvb2YiOnsicHJvb2ZQdXJwb3NlIjoicHJvb2ZQdXJwb3NlIiwidHlwZSI6Ikpzb25XZWJTaWduYXR1cmUyMDIwIiwidmVyaWZpY2F0aW9uTWV0aG9kIjoiZGlkOndlYjpsb2NhbGhvc3Q6QlBOTDAwMDAwMDAwMDAwMCMiLCJjcmVhdGVkIjoiMjAyMy0wNy0xOVQwOToxMTozOVoiLCJqd3MiOiJleUpoYkdjaU9pSkZaRVJUUVNKOS4uZmRuMnFVODVhdU9sdGRIRExkSEk3c0pWVjFaUGRmdHBpWGRfbmRYTjBkRmdTRFdpSXJTY2REMDN3dHZLTHFfSC1zaFFXZmgyUlllTW1ybEV6QWhmRHcifX19LCJleHAiOjE2ODk3NTg1NDQsImp0aSI6IjdlNWE4MzQ4LTgwZjUtNGIzMS1iMDNlLTBiOTJmNzc4ZTVjZiJ9.c7FS-CLwm3vxfO9847M5sqcVxv3QbwwSmSsFWcGif7MOesjt1pdnARlQ4pvHzgsFj1UqBEvHwZQvyYyPCQg_Cw"
                  }
                },
                "VP as json-ld": {
                  "description": "VP as json-ld",
                  "value": {
                    "vp": {
                      "id": "b9d97cef-758d-4a7c-843d-86f17632b08a",
                      "type": [
                        "VerifiablePresentation"
                      ],
                      "@context": [
                        "https://www.w3.org/2018/credentials/v1"
                      ],
                      "verifiableCredential": [
                        {
                          "id": "did:web:localhost:BPNL000000000000#f73e3631-ba87-4a03-bea3-b28700056879",
                          "@context": [
                            "https://www.w3.org/2018/credentials/v1",
                            "https://catenax-ng.github.io/product-core-schemas/businessPartnerData.json",
                            "https://w3id.org/security/suites/jws-2020/v1"
                          ],
                          "type": [
                            "VerifiableCredential",
                            "BpnCredential"
                          ],
                          "issuer": "did:web:localhost:BPNL000000000000",
                          "expirationDate": "2024-12-31T18:30:00Z",
                          "issuanceDate": "2023-07-19T09:11:34Z",
                          "credentialSubject": [
                            {
                              "bpn": "BPNL000000000000",
                              "id": "did:web:localhost:BPNL000000000000",
                              "type": "BpnCredential"
                            }
                          ],
                          "proof": {
                            "created": "2023-07-19T09:11:39Z",
                            "jws": "eyJhbGciOiJFZERTQSJ9..fdn2qU85auOltdHDLdHI7sJVV1ZPdftpiXd_ndXN0dFgSDWiIrScdD03wtvKLq_H-shQWfh2RYeMmrlEzAhfDw",
                            "proofPurpose": "proofPurpose",
                            "type": "JsonWebSignature2020",
                            "verificationMethod": "did:web:localhost:BPNL000000000000#"
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Verifiable presentation validate",
            "content": {
              "application/json": {
                "examples": {
                  "VP as JWT": {
                    "description": "VP as JWT",
                    "value": {
                      "valid": true,
                      "validateJWTExpiryDate": true,
                      "validateAudience": true,
                      "vp": "eyJraWQiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwIiwidHlwIjoiSldUIiwiYWxnIjoiRWREU0EifQ.eyJzdWIiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwIiwiYXVkIjoic21hcnQiLCJpc3MiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwIiwidnAiOnsiaWQiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwIzM4ZTU2ZTg1LTNkODQtNGEyNS1iZjg1LWFiMjRlYzY4MmMwOSIsInR5cGUiOlsiVmVyaWZpYWJsZVByZXNlbnRhdGlvbiJdLCJAY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvMjAxOC9jcmVkZW50aWFscy92MSJdLCJ2ZXJpZmlhYmxlQ3JlZGVudGlhbCI6eyJAY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvMjAxOC9jcmVkZW50aWFscy92MSIsImh0dHBzOi8vY2F0ZW5heC1uZy5naXRodWIuaW8vcHJvZHVjdC1jb3JlLXNjaGVtYXMvYnVzaW5lc3NQYXJ0bmVyRGF0YS5qc29uIiwiaHR0cHM6Ly93M2lkLm9yZy9zZWN1cml0eS9zdWl0ZXMvandzLTIwMjAvdjEiXSwidHlwZSI6WyJWZXJpZmlhYmxlQ3JlZGVudGlhbCIsIkJwbkNyZWRlbnRpYWwiXSwiaWQiOiJkaWQ6d2ViOmxvY2FsaG9zdDpCUE5MMDAwMDAwMDAwMDAwI2Y3M2UzNjMxLWJhODctNGEwMy1iZWEzLWIyODcwMDA1Njg3OSIsImlzc3VlciI6ImRpZDp3ZWI6bG9jYWxob3N0OkJQTkwwMDAwMDAwMDAwMDAiLCJpc3N1YW5jZURhdGUiOiIyMDIzLTA3LTE5VDA5OjExOjM0WiIsImV4cGlyYXRpb25EYXRlIjoiMjAyNC0xMi0zMVQxODozMDowMFoiLCJjcmVkZW50aWFsU3ViamVjdCI6eyJpZCI6ImRpZDp3ZWI6bG9jYWxob3N0OkJQTkwwMDAwMDAwMDAwMDAiLCJicG4iOiJCUE5MMDAwMDAwMDAwMDAwIiwidHlwZSI6IkJwbkNyZWRlbnRpYWwifSwicHJvb2YiOnsicHJvb2ZQdXJwb3NlIjoicHJvb2ZQdXJwb3NlIiwidHlwZSI6Ikpzb25XZWJTaWduYXR1cmUyMDIwIiwidmVyaWZpY2F0aW9uTWV0aG9kIjoiZGlkOndlYjpsb2NhbGhvc3Q6QlBOTDAwMDAwMDAwMDAwMCMiLCJjcmVhdGVkIjoiMjAyMy0wNy0xOVQwOToxMTozOVoiLCJqd3MiOiJleUpoYkdjaU9pSkZaRVJUUVNKOS4uZmRuMnFVODVhdU9sdGRIRExkSEk3c0pWVjFaUGRmdHBpWGRfbmRYTjBkRmdTRFdpSXJTY2REMDN3dHZLTHFfSC1zaFFXZmgyUlllTW1ybEV6QWhmRHcifX19LCJleHAiOjE2ODk4MzQ4MDUsImp0aSI6ImIwODYzOWZiLWQ5MWEtNGUwZS1iNmY4LTYzYjdhMzQ1ZTRhZiJ9.80x0AB-OauefdeZfx1cwhitdVKRvCRFeFzYwU73DL7y4w34vu6BdfHWLBGjkwELxkQEoFfiTPOqtuyqhtsyDBg",
                      "validateExpiryDate": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The input does not comply to the syntax requirements",
            "content": {
              "application/json": {
                "examples": {
                  "Validation of VP in form of JSON-LD is not supported": {
                    "description": "Validation of VP in form of JSON-LD is not supported",
                    "value": {
                      "type": "about:blank",
                      "title": "Validation of VP in form of JSON-LD is not supported",
                      "status": 400,
                      "detail": "Validation of VP in form of JSON-LD is not supported",
                      "instance": "/api/presentations/validation",
                      "properties": {
                        "timestamp": 1689835085703
                      }
                    }
                  },
                  "Response in case of invalid data provided": {
                    "description": "Response in case of invalid data provided",
                    "value": {
                      "type": "about:blank",
                      "title": "Invalid data provided",
                      "status": 400,
                      "detail": "details",
                      "instance": "API endpoint",
                      "properties": {
                        "timestamp": 1689760833962,
                        "errors": {
                          "filed": "filed error message"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "The request could not be completed due to a failed authorization."
          },
          "403": {
            "description": "The request could not be completed due to a forbidden access"
          },
          "500": {
            "description": "Any other internal server error",
            "content": {
              "application/json": {
                "examples": {
                  "Internal server error": {
                    "description": "Internal server error",
                    "value": {
                      "type": "about:blank",
                      "title": "Error Title",
                      "status": 500,
                      "detail": "Error Details",
                      "instance": "API endpoint",
                      "properties": {
                        "timestamp": 1689762476720
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authenticate using access_token": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "Authenticate using access_token": {
        "type": "apiKey",
        "description": "**Bearer (apiKey)**\nJWT Authorization header using the Bearer scheme.\nEnter **Bearer** [space] and then your token in the text input below:\nExample: Bearer 12345abcdef\n",
        "name": "Authorization",
        "in": "header"
      },
      "sts_token": {
        "type": "apiKey",
        "description": "**STS token**\nJWT Authorization header.\nEnter your token in the text input below:\nExample: 12345abcdef\n",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}